+++ app/admin/background_jobs.rb
@@ -45,6 +45,7 @@ ActiveAdmin.register BackgroundJob do
       end
     end
     column :completed_at
+    column :user
 
     customizable_default_actions(current_ability) do |background_job|
       background_job.finished? ? [] : [:destroy]